fixup cherry-pick mistake
